formatting issues

+++ src/main/java/edu/luc/cs271/myhashmap/Main.java
@@ -8,7 +8,7 @@ public class Main {
     // set up the scanner so that it separates words based on space and punctuation
     final Scanner input = new Scanner(System.in).useDelimiter("[^\\p{Alnum}]+");
     // DONE?TODO measure the performance for MyHashMap, HashMap, and TreeMap several times each!
-    int begintime = System.currentTimeMillis();
+    long begintime = System.currentTimeMillis();
     final Map<String, Integer> counts = new MyHashMap<>(6007); // a prime number!
     final long time0 = System.currentTimeMillis(); // current time
     while (input.hasNext()) {
@@ -23,7 +23,7 @@ public class Main {
     Collections.sort(arr, new DescendingByCount());
     for (int i = 0; i < 10 && i < arr.size(); i += 1) {
       System.out.println(arr.get(i));
-      int endtime = System.currentTimeMillis();
+      long endtime = System.currentTimeMillis();
       System.out.println(endtime - begintime);
     }
   }
